[#2815] Enhancement StringUtils  - Add isNotEmpty()  - Move from bootstrap-core module to commons module    - Deprecate StringUtils of bootstrap-core module  - Change splitAndTrim() API  - splitAndTrim() -> tokenizeToStringList()     - Copy from spring framework
